requirements.txt
attrs==20.3.0
botocore==1.19.56
chalice==1.21.9
click==7.1.2
enum-compat==0.0.3
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
Pillow==8.1.0
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.15.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
